Create render_stdin.py to render lines etc
Python script to use Tkinter (TK, TCL) on X11 to render lines etc on the command line.  The goal is that we can manually verify the geometry library is working as expected by looking at lines and their intersections and verifying they look correct.  So we will write a driver that calls the geometry library and outputs stuff to stdout to render (e.g. the lines and where we think the intersection is)